multiple arch builders on circleci
