+++ .circleci/config.yml
@@ -1,15 +1,29 @@
 executors:
-  latest-image:
+  amd64-large-executor:
     machine:
       image: ubuntu-2004:202111-02
       docker_layer_caching: true
+      resource_class: large
+  amd64-medium-executor:
+    machine:
+      image: ubuntu-2004:202111-02
+      docker_layer_caching: true
+      resource_class: medium
+  amd64-small-executor:
+    machine:
+      image: ubuntu-2004:202111-02
+      docker_layer_caching: true
+      resource_class: small
+  arm64-medium-executor:
+    machine:
+      image: ubuntu-2004:202101-01
+      docker_layer_caching: true
+      resource_class: arm.medium
 jobs:
-  build:
-    executor: latest-image
+  build-base-amd64:
+    executor: amd64-medium-executor
     steps:
       - checkout
-      - restore_cache:
-          key: v1-{{ .Revision }}
       - run:
           command: ./docker/docker-login.sh
           name: Login to Docker Hub
@@ -22,6 +36,39 @@ jobs:
       - run:
           command: ./docker-build-codex-base.sh
           name: Build Base Image
+  build-base-arm64:
+    executor: arm64-medium-executor
+    steps:
+      - checkout
+      - run:
+          command: ./docker/docker-login.sh
+          name: Login to Docker Hub
+      - run:
+          command: ./docker/docker-install-binfmt-platforms.sh
+          name: Install binfmt platforms
+      - run:
+          command: ./docker/docker-buildx-create.sh
+          name: Create buildx builder
+      - run:
+          command: ./docker-build-codex-base.sh
+          name: Build Base Image
+  build-dist:
+    executor: amd64-large-executor
+    steps:
+      - checkout
+      - restore_cache:
+          key: wheels-v1-{{ arch }}
+      - restore_cache:
+          key: dist-v1-{{ .Revision }}
+      - run:
+          command: ./docker/docker-login.sh
+          name: Login to Docker Hub
+      - run:
+          command: ./docker/docker-install-binfmt-platforms.sh
+          name: Install binfmt platforms
+      - run:
+          command: ./docker/docker-buildx-create.sh
+          name: Create buildx builder
       - run:
           command: ./docker-build-codex-builder-base.sh
           name: Build Builder Base Image
@@ -29,8 +76,12 @@ jobs:
           command: ./docker-build-codex-builder.sh
           name: Build Builder Image
       - run:
-          command: ./docker-build-codex-wheels.sh
-          name: Build Wheels Cache Image
+          command: ./docker/docker-compose-exit.sh codex-save-cache
+          name: Save python caches to host
+      - save_cache:
+          key: wheels-v1-{{ arch }}
+          paths:
+            - ./cache
       - run:
           command: ./docker/docker-compose-exit.sh codex-frontend-lint
           name: "Frontend: Lint"
@@ -58,27 +109,23 @@ jobs:
           command: ./docker/docker-compose-exit.sh codex-build-dist
           name: Build Distribution
       - save_cache:
-          key: v1-{{ .Revision }}
+          key: dist-cache-v1-{{ .Revision }}
           paths:
             - dist
       - persist_to_workspace:
           paths:
-            - ./.env
-            - ./.env.platforms
-            - ./Dockerfile
-            - ./builder-requirements.txt
             - ./dist
-            - ./docker
-            - ./docker-build-codex.sh
-            - ./pypi-deploy.sh
           root: .
-  deploy:
-    executor: latest-image
+  build-amd64:
+    executor: amd64-medium-executor
     steps:
+      - checkout
       - attach_workspace:
           at: .
       - restore_cache:
-          key: v1-{{ .Revision }}
+          key: dist-cache-v1-{{ .Revision }}
+      - restore_cache:
+          key: wheels-cache-v1-{{ arch }}
       - run:
           command: ./docker/docker-login.sh
           name: Login to Docker Hub
@@ -90,7 +137,36 @@ jobs:
           name: Create buildx builder
       - run:
           command: ./docker-build-codex.sh
-          name: Build & Push Codex Runnable Image
+          name: Build Codex Runnable Image
+  build-arm64:
+    executor: arm64-medium-executor
+    steps:
+      - attach_workspace:
+          at: .
+      - restore_cache:
+          key: dist-cache-v1-{{ .Revision }}
+      - restore_cache:
+          key: wheels-cache-v1-{{ arch }}
+      - run:
+          command: ./docker/docker-login.sh
+          name: Login to Docker Hub
+      - run:
+          command: ./docker/docker-install-binfmt-platforms.sh
+          name: Install binfmt platforms
+      - run:
+          command: ./docker/docker-buildx-create.sh
+          name: Create buildx builder
+      - run:
+          command: ./docker-build-codex.sh
+          name: Build Codex Runnable Image
+  deploy:
+    executor: amd64-small-executor
+    steps:
+      - attach_workspace:
+          at: .
+      - run:
+          command: ./docker-create-multiarch-codex.sh
+          name: Create multi-arch images
       - run:
           command: ./pypi-deploy.sh
           name: Publish Codex Package to PyPi
@@ -98,18 +174,51 @@ version: 2.1
 workflows:
   main:
     jobs:
-      - build:
+      - build-base-amd64:
           filters:
             branches:
               only:
                 - develop
                 - release
                 - /.*-ci/
+      - build-base-arm64:
+          filters:
+            branches:
+              only:
+                - develop
+                - release
+                - /.*-ci/
+      - build-dist:
+          filters:
+            branches:
+              only:
+                - develop
+                - release
+                - /.*-ci/
+          requires:
+            - build-base-amd64
+      - build-amd64:
+          filters:
+            branches:
+              only:
+                - release
+          requires:
+            - build-dist
+            - build-base-amd64
+      - build-arm64:
+          filters:
+            branches:
+              only:
+                - release
+          requires:
+            - build-dist
+            - build-base-arm64
       - deploy:
           filters:
             branches:
               only:
                 - release
           requires:
-            - build
+            - build-amd64
+            - build-arm64
   version: 2.1
