+++ docker-version-codex-base.sh
@@ -10,7 +10,11 @@ DEPS=(
     docker-build-codex-base.sh
 )
 DEPS_MD5S=$(md5sum "${DEPS[@]}")
-echo -e "$PYTHON_ALPINE_VERSION  python-alpine-version\n$DEPS_MD5S" |
+VERSION=$(echo -e "$PYTHON_ALPINE_VERSION  python-alpine-version\n$DEPS_MD5S" |
     LC_ALL=C sort |
     md5sum |
-    awk '{print $1}'
+    awk '{print $1}')
+if [[ ${CIRCLECI:-} ]]; then
+    VERSION="${VERSION}-$(uname -m)"
+fi
+echo "$VERSION"
